move extraction code into its own component
