+++ BUILD.bazel
@@ -11,6 +11,7 @@ cc_library(
     "src/datadog/default_http_client_null.cpp",
     "src/datadog/environment.cpp",
     "src/datadog/error.cpp",
+    "src/datadog/extraction_util.cpp",
     "src/datadog/glob.cpp",
     "src/datadog/id_generator.cpp",
     "src/datadog/limiter.cpp",
@@ -59,6 +60,7 @@ cc_library(
     "src/datadog/event_scheduler.h",
     "src/datadog/expected.h",
     "src/datadog/extracted_data.h",
+    "src/datadog/extraction_util.h",
     "src/datadog/glob.h",
     "src/datadog/hex.h",
     "src/datadog/http_client.h",
